TagReader replaced by TagManager which allows not only read but also to write to NBTTag
