+++ pom.xml
@@ -4,7 +4,7 @@
 
   <groupId>vg.civcraft.mc.civmodcore</groupId>
   <artifactId>CivModCore</artifactId>
-  <version>1.5.06</version>
+  <version>1.5.07</version>
 
   <name>CivModCore</name>
   <url>https://github.com/Civcraft/CivModCore/</url>
